// src/api/models/homeResponseDto.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * YRC Race API
 * 여의도 러닝 크루 마라톤 기록 API
 * OpenAPI spec version: v1.0
 */
import type { BestRecordInfo } from './bestRecordInfo';
import type { RaceInfoResponseDto } from './raceInfoResponseDto';
import type { RecordInfoResponseDto } from './recordInfoResponseDto';

export interface HomeResponseDto {
  /** 사용자 id */
  memberId?: number;
  /** 사용자 이름 */
  memberName?: string;
  /** full 최고 기록 */
  bestFullRecord?: BestRecordInfo;
  /** half 최고 기록 */
  bestHalfRecord?: BestRecordInfo;
  /** 10K 최고 기록 */
  bestTenRecord?: BestRecordInfo;
  /** 참여 가능한 대회 정보 */
  raceInfos?: RaceInfoResponseDto[];
  /** 기록 확인 가능한 대회 정보 */
  recordInfos?: RecordInfoResponseDto[];
}